관심상품 카운트 구현 성공, create/delete 구현 성공

+++ shoekream/src/main/java/com/supreme/shoekream/model/dto/ProductDTO.java
@@ -11,7 +11,7 @@ public record ProductDTO(
         String name,
         String nameKor,
         String size,
-        Long wishCount,
+        int wishCount,
         String modelNum,
         String releaseDate,
         String color,
@@ -21,7 +21,7 @@ public record ProductDTO(
         String collection
 ) {
 
-    public static ProductDTO of(Long idx, String img, String brand, String name, String nameKor, String size, Long wishCount, String modelNum, String releaseDate, String color, String firstPrice, String category, String gender, String collection ){
+    public static ProductDTO of(Long idx, String img, String brand, String name, String nameKor, String size, int wishCount, String modelNum, String releaseDate, String color, String firstPrice, String category, String gender, String collection ){
         return new ProductDTO(idx, img, brand, name, nameKor, size, wishCount, modelNum, releaseDate, color, firstPrice, category, gender, collection);
     }
